rectify install path for toolbox install
INSTALL_PATH is now reset after the platform check is finished, creating
the missing folder with it if need be. This also keeps the INSTALL_PATH
changes isolated to the top section, leaving the install_cmd to smaller set of things to do.

the command was not run for the container environment thus
not adding links to the tools anymore so this introduces it back

finally, to avoid adding multiple times the install path in the PATH a
quick grep check on the profile will limit runs to only once